spawn: port Spawn_Splash to TRX
