+++ src/libtrx/game/spawn.c
@@ -1,0 +1,29 @@
+#include "game/spawn.h"
+
+#include "game/effects.h"
+#include "game/random.h"
+#include "game/rooms.h"
+
+void Spawn_Splash(const ITEM *const item)
+{
+    const int32_t water_height = Room_GetWaterHeight(
+        item->pos.x, item->pos.y, item->pos.z, item->room_num);
+    int16_t room_num = item->room_num;
+    Room_GetSector(item->pos.x, item->pos.y, item->pos.z, &room_num);
+
+    for (int32_t i = 0; i < 10; i++) {
+        const int16_t effect_num = Effect_Create(room_num);
+        if (effect_num == NO_EFFECT) {
+            continue;
+        }
+
+        EFFECT *const effect = Effect_Get(effect_num);
+        effect->object_id = O_SPLASH_1;
+        effect->pos.x = item->pos.x;
+        effect->pos.y = water_height;
+        effect->pos.z = item->pos.z;
+        effect->rot.y = 2 * Random_GetDraw() + DEG_180;
+        effect->speed = Random_GetDraw() / 256;
+        effect->frame_num = 0;
+    }
+}
